docs: Write some documentation on structs
